# api/openapi.yaml
openapi: 3.0.3
info:
  title: QuizMind app API
  version: 0.1.0
  description: |
    Endpoints used by the Flutter app. The admin API (/admin/*) is consumed only by the
    bundled Vue UI and is not part of this contract.
    These endpoints need no authentication; only the admin API (/admin/*) is guarded by the server token.
    Timestamps are Unix milliseconds. Errors are `{"error": "message"}`.
servers:
  - url: http://127.0.0.1:8080
paths:
  /healthz:
    get:
      summary: Liveness probe
      responses:
        "200": { description: ok }
  /api/v1/banks:
    get:
      summary: Banks with their published question count
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Bank" }
  /api/v1/sync/questions:
    get:
      summary: Incremental question download
      description: |
        Returns changes with sync_seq > since in sync_seq order. Published questions come back in
        `items`; questions that are no longer published come back in `deleted`. Repeat with
        `since = next_seq` while `has_more` is true.
      parameters:
        - { name: since, in: query, schema: { type: integer, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, default: 500, maximum: 1000 } }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuestionsPage" }
  /api/v1/sync/states:
    get:
      summary: Incremental download of per-question learning state
      parameters:
        - { name: since, in: query, schema: { type: integer, default: 0 }, description: "server sync_seq cursor" }
        - { name: limit, in: query, schema: { type: integer, default: 500, maximum: 1000 } }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatesPage" }
    post:
      summary: Upload states (last writer wins on updated_at)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 500
              items: { $ref: "#/components/schemas/State" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadResult" }
  /api/v1/sync/sessions:
    get:
      summary: Incremental download of saved quizzes ("continue where you left off")
      parameters:
        - { name: since, in: query, schema: { type: integer, default: 0 }, description: "server sync_seq cursor" }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 100 } }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionsPage" }
    post:
      summary: Upload saved quizzes (last writer wins on updated_at)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 50
              items: { $ref: "#/components/schemas/SessionProgress" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadResult" }
        "400": { description: invalid session }
  /api/v1/sync/attempts:
    get:
      summary: Incremental download of the answer log (so every device sees the whole history)
      parameters:
        - { name: since, in: query, schema: { type: integer, default: 0 }, description: "server sync_seq cursor" }
        - { name: limit, in: query, schema: { type: integer, default: 500, maximum: 1000 } }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttemptsPage" }
    post:
      summary: Append answer records (idempotent on id)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 500
              items: { $ref: "#/components/schemas/Attempt" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadResult" }
        "400": { description: invalid attempt }
  /api/v1/sync/exams:
    get:
      summary: Incremental download of finished mock exams
      parameters:
        - { name: since, in: query, schema: { type: integer, default: 0 }, description: "server sync_seq cursor" }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 100 } }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExamsPage" }
    post:
      summary: Upload finished exams (append-only, idempotent on id)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 10
              items: { $ref: "#/components/schemas/Exam" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadResult" }
        "400": { description: invalid exam }
  /api/v1/sync/notes:
    get:
      summary: Incremental download of AI explanations (one per question)
      parameters:
        - { name: since, in: query, schema: { type: integer, default: 0 }, description: "server sync_seq cursor" }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 100 } }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotesPage" }
    post:
      summary: Upload AI explanations (last writer wins on updated_at; unknown questions are ignored)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 100
              items: { $ref: "#/components/schemas/AiNote" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UploadResult" }
        "400": { description: invalid note }
  /api/v1/ai/config:
    get:
      summary: The OpenAI-compatible endpoint the app calls for AI explanations (includes the API key)
      description: >
        Configured in the admin UI. Unlike the rest of /api/v1 this needs the access token
        set there: `Authorization: Bearer <token>`. 404 means the feature is switched off,
        and the app should drop its copy of the configuration.
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiConfig" }
        "401": { description: missing or wrong access token }
        "404": { description: AI explanations are not enabled }
  /api/v1/questions/{id}/flag:
    post:
      summary: Report that a question has a problem; two unresolved reports take it offline
      description: |
        The body is optional. Clients that send none (or an unknown reason) are recorded as `other`.
        A reviewer settling the reports (approve, reject, or "dismiss flags" in the admin UI) resets
        the count, so `flag_count` is the number of unresolved reports.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  enum: [wrong_answer, ambiguous, typo, other]
                  description: wrong answer / ambiguous stem / typo in options or text / other
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  flag_count: { type: integer }
                  status: { type: string }
        "404": { description: unknown question }
components:
  schemas:
    Bank:
      type: object
      required: [id, title, description, question_count]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        question_count: { type: integer }
    Question:
      type: object
      required: [id, bank_id, type, stem, options, answer, explanation, difficulty, tags, source_quote, sync_seq, updated_at]
      properties:
        id: { type: string }
        bank_id: { type: string }
        type: { type: string, enum: [single, judge] }
        stem: { type: string }
        options: { type: array, items: { type: string } }
        answer: { type: array, items: { type: integer }, description: "option indexes; one entry for single and judge" }
        explanation: { type: string }
        difficulty: { type: integer, minimum: 1, maximum: 5 }
        tags: { type: array, items: { type: string } }
        source_quote: { type: string }
        sync_seq: { type: integer }
        updated_at: { type: integer }
    QuestionsPage:
      type: object
      required: [items, deleted, next_seq, has_more]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Question" } }
        deleted: { type: array, items: { type: string } }
        next_seq: { type: integer }
        has_more: { type: boolean }
    State:
      type: object
      required: [question_id, favorite, wrong_count, updated_at]
      properties:
        question_id: { type: string }
        fsrs: { type: object, nullable: true, description: "opaque to the server" }
        due_at: { type: integer, nullable: true }
        favorite: { type: boolean }
        wrong_count: { type: integer }
        updated_at: { type: integer }
        sync_seq: { type: integer, readOnly: true }
    StatesPage:
      type: object
      required: [items, next_seq, has_more]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/State" } }
        next_seq: { type: integer }
        has_more: { type: boolean }
    Attempt:
      type: object
      required: [id, question_id, device_id, answer, is_correct, answered_at]
      properties:
        id: { type: string, description: "client-generated ULID" }
        question_id: { type: string }
        device_id: { type: string }
        answer: { type: array, items: { type: integer } }
        is_correct: { type: boolean }
        duration_ms: { type: integer, nullable: true, description: "time from showing the question to answering it" }
        review_ms: { type: integer, nullable: true, description: "time spent on the question after answering (explanation, AI); re-uploading with a larger value raises it, a smaller one is ignored" }
        answered_at: { type: integer }
    AttemptsPage:
      type: object
      required: [items, next_seq, has_more]
      properties:
        items:
          type: array
          items:
            allOf:
              - { $ref: "#/components/schemas/Attempt" }
              - type: object
                properties:
                  sync_seq: { type: integer, readOnly: true }
        next_seq: { type: integer }
        has_more: { type: boolean }
    Exam:
      type: object
      required: [id, bank_id, finished_at, total, correct, answered, percent, passed, used_ms, items]
      description: |
        A finished mock exam. It never changes after it is handed in. `items` lists the
        paper in order so any device can review it.
      properties:
        id: { type: string, description: "client-generated ULID" }
        bank_id: { type: string }
        title: { type: string, maxLength: 200 }
        finished_at: { type: integer }
        total: { type: integer }
        correct: { type: integer }
        answered: { type: integer, description: "questions that got an answer; the rest were left blank" }
        percent: { type: integer, minimum: 0, maximum: 100 }
        passed: { type: boolean }
        limit_sec: { type: integer, nullable: true, description: "time allowed; null when untimed" }
        used_ms: { type: integer }
        device_id: { type: string, maxLength: 64 }
        sync_seq: { type: integer, readOnly: true }
        items:
          type: array
          maxItems: 1000
          items:
            type: object
            required: [q, s, c]
            properties:
              q: { type: string, description: "question id" }
              s: { type: array, items: { type: integer }, description: "option indexes picked; empty = left blank" }
              c: { type: boolean, description: "answered correctly" }
    ExamsPage:
      type: object
      required: [items, next_seq, has_more]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Exam" } }
        next_seq: { type: integer }
        has_more: { type: boolean }
    SessionProgress:
      type: object
      required: [scope, updated_at]
      description: |
        The quiz one device left part-way in one scope (a bank id), so another device can
        continue it. There is at most one per scope. A finished quiz is uploaded with
        `data: null`; that tombstone clears the saved quiz on every device.
      properties:
        scope: { type: string, maxLength: 64, description: "bank id" }
        data: { allOf: [{ $ref: "#/components/schemas/SessionData" }], nullable: true }
        updated_at: { type: integer, description: "client clock, ms; the newer one wins" }
        device_id: { type: string, maxLength: 64 }
        sync_seq: { type: integer, readOnly: true }
    SessionData:
      type: object
      description: |
        Opaque to the server (at most 512 KiB); every client reads and writes this shape.
        Questions and answers are kept by id, so questions withdrawn since are simply skipped.
      required: [title, ids, seed, index, answers]
      properties:
        title: { type: string }
        ids: { type: array, items: { type: string }, description: "question ids in the order they are practised" }
        seed: { type: integer, description: "seeds the option shuffle so a resumed quiz shows the same layout" }
        index: { type: integer, description: "position in ids of the current question" }
        answers:
          type: object
          description: "by question id: s = chosen option indexes (original order), c = whether it was right"
          additionalProperties:
            type: object
            required: [s, c]
            properties:
              s: { type: array, items: { type: integer } }
              c: { type: boolean }
    SessionsPage:
      type: object
      required: [items, next_seq, has_more]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/SessionProgress" } }
        next_seq: { type: integer }
        has_more: { type: boolean }
    AiConfig:
      type: object
      required: [base_url, api_key, model, max_tokens, temperature]
      properties:
        base_url: { type: string, description: "including the version segment; the app appends /chat/completions" }
        api_key: { type: string }
        model: { type: string }
        max_tokens: { type: integer }
        temperature: { type: number }
    AiNote:
      type: object
      required: [question_id, content, updated_at]
      properties:
        question_id: { type: string }
        content: { type: string, description: "Markdown, at most 64 KB" }
        model: { type: string }
        prompt_version: { type: string }
        selected: { type: array, items: { type: integer }, description: "option indexes the learner had picked when asking" }
        updated_at: { type: integer, description: "client clock, ms; last writer wins" }
        device_id: { type: string }
        sync_seq: { type: integer, description: "server-assigned; present on download" }
    NotesPage:
      type: object
      required: [items, next_seq, has_more]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AiNote" } }
        next_seq: { type: integer }
        has_more: { type: boolean }
    UploadResult:
      type: object
      properties:
        accepted: { type: integer }
        ignored: { type: integer }